Allow title/subtitle labels to size to fit
Normally the title/subtitle labels expand to use all the available size
of the the superview so text alignment works as expected. In cases where
text alignment will not be used the title/subtitle labels can be set to
size to fit so they only take up as much width as needed.
This functionality is toggleable via the titleSubtitleLabelsSizeToFit
apperance proxy property. Note: By default when enabling this feature
elements will expand as needed from the center horizontally.