Prevent re-setting of titleSubtitleLabelsSizeToFit
This fixes a crash due to a UIAppearance bug where UI_APPEARANCE_SELECTOR
methods are called as many times as they are set for the class. The
titleSubtitleLabelsSizeToFit property did/does not expect to be set
multiple times in a row hence: crash.

UIAppareance bug documentation can be found here:
http://www.openradar.me/28827675