Fix presentation regression on iOS 11 non iPhoneX devices
Fix bug introduced in commit dacab111c that improperly offset
presentation of RMessage on non notch devices running iOS 11.

Note: Normally when presenting RMessage on non-notch devices we offset the title 10pts
from the top or bottom depending on presentation. When presenting on notch
devices there seems to be padding in the safe area so we tweak presentation
here to not offset by the normal 10pts so as to maintain some visual consistency.