Setup view layout right before animation
Previous to this commit RMessageView would setup various layout info
(constraints etc..) and store this layout information in its instance.
So for example when queueing 10 messages to be displayed layout information would be
stored in the RMessageView class instances for displaying/manipulating/animating later.
This behavior (obviously) would lead to crashes if layout information changed
between when the RMessageView would be instantiated and when it was
presented.

This commit changes the above behavior so that RMessageView
no longer stores layout information relating its superview (which
can change between instantiation and presentation) on instantiation.
Instead this layout information is delayed to be processed until right
before the instance of the RMessageView is told to animate which is
the best time for layout information NOT to change and hence lead to crashes.