Fix: Size RMessageView according to layoutGuide
Stop the super hacky sizing of RMessageView based on detection of the
parent view controller. It seems this was carried over from TSMessages
prior to layoutGuide support being in iOS. Layout guide was used instead
of safe areas for compatibility with iOS 7 devices. Release 3.x.x will
highly likely be the last version of RMessage to support iOS7.