Fix sizing issue on iOS versions less than 11.0
Issue: UITableViewControllers in iOS versions below 11 offset the entire
topLayoutGuide downward when the automaticallyAdjustsScrollViewInsets property
is set (which it is by default). This causes the topLayoutGuide bottom which
normally is located at the bottom of any top bars to now be offset
(height of top bar) pts below the bottom of the bar. This is not expected
behavior (most likely a bug) and causes RMessage to improperly size itself.

This commit fixes the said issue by detecting the presence of the
automaticallyAdjustsScrollViewInsets being set to true, iOS 10 running,
and RMessage's superview being a scrollview.

https://github.com/donileo/RMessage/issues/53